refactor(store-core): extract session-lifecycle handlers into handlers/session-lifecycle.ts (audit P2-3)
Move the session-lifecycle family (handleSessionUpdated, handleSessionError +
its boundSessionMismatchMessage helper, handleSessionStopped, handleLogEntry +
LogLevel/LogEntry/VALID_LOG_LEVELS) out of the handlers/index.ts god-module into
handlers/session-lifecycle.ts, behind the barrel's export * re-export so the
public surface is unchanged.

handleError and its private helpers (DEFAULT_ERROR_MESSAGE, pickFiniteTokenCount)
stay in index.ts for the error-family slice. Drop the now-unused SessionInfo type
import from index.ts (it was only used by handleSessionUpdated).

Pure move — bodies byte-identical to origin/main. index.ts 2572 -> 2362 lines.
